TINKERPOP-2671 Added tx() support in gremlin-language
Adding tx() support to gremlin-language brings its grammar/parser closer in line with what is possible now with bytecode and with Groovy scripts. From the user's perspective there is now a clear way for them to signify that their transaction is closed with either rollback or commit. Support for begin() is present for consistency but perhaps it shouldn't be in the context of scripts as it is meant to spawn a Transaction instance and since you can't assign variables in gremlin-language that doesn't seem useful.